@@ -5,24 +5,22 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>3.3.6</version>
+        <version>4.0.1</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
 
     <groupId>cz.gattserver</groupId>
     <artifactId>grass4</artifactId>
-    <version>4.0.2</version>
+    <version>4.0.3</version>
     <name>grass4</name>
     <description>GRASS4 Knowledge Base</description>
     <packaging>jar</packaging>
 
     <properties>
-        <java.version>17</java.version>
-        <vaadin.version>24.5.5</vaadin.version>
+        <java.version>21</java.version>
+        <vaadin.version>25.0.3</vaadin.version>
         <tika.core.version>3.0.0</tika.core.version>
         <org.mapstruct.version>1.6.3</org.mapstruct.version>
-        <maven-compiler-plugin-version>3.9.0</maven-compiler-plugin-version>
-        <querydsl-apt.version>5.1.0</querydsl-apt.version>
         <h2.version>2.1.214</h2.version>
     </properties>
 
@@ -59,6 +57,11 @@
             <!-- Replace artifactId with vaadin-core to use only free components -->
             <!-- <artifactId>vaadin</artifactId> -->
             <artifactId>vaadin-core</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.vaadin</groupId>
+            <artifactId>vaadin-dev</artifactId>
+            <optional>true</optional>
         </dependency>
         <dependency>
             <groupId>com.vaadin</groupId>
@@ -217,12 +220,6 @@
             <version>${querydsl.version}</version>
             <classifier>jakarta</classifier>
         </dependency>
-        <dependency>
-            <groupId>com.querydsl</groupId>
-            <artifactId>querydsl-apt</artifactId>
-            <version>${querydsl.version}</version>
-            <classifier>jakarta</classifier>
-        </dependency>
 
         <!-- Tomcat JDBC -->
         <dependency>
@@ -324,6 +321,30 @@
                 </configuration>
             </plugin>
 
+
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <version>3.14.1</version>
+                <configuration>
+                    <source>${java.version}</source>
+                    <target>${java.version}</target>
+
+                    <annotationProcessorPaths>
+                        <path>
+                            <groupId>com.querydsl</groupId>
+                            <artifactId>querydsl-apt</artifactId>
+                            <version>${querydsl.version}</version>
+                            <classifier>jakarta</classifier>
+                        </path>
+                        <path>
+                            <groupId>jakarta.persistence</groupId>
+                            <artifactId>jakarta.persistence-api</artifactId>
+                            <version>3.1.0</version>
+                        </path>
+                    </annotationProcessorPaths>
+                </configuration>
+            </plugin>
         </plugins>
     </build>
 
@@ -342,6 +363,7 @@
                 </resources>
             </build>
         </profile>
+
         <profile>
             <!-- Production mode is activated using -Pproduction -->
             <id>prod</id>
