@@ -15,7 +15,6 @@ import org.springframework.security.web.context.DelegatingSecurityContextReposit
 import org.springframework.security.web.context.HttpSessionSecurityContextRepository;
 import org.springframework.security.web.context.RequestAttributeSecurityContextRepository;
 import org.springframework.security.web.context.SecurityContextRepository;
-import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
 
 import java.util.List;
 
@@ -25,45 +24,45 @@ import java.util.List;
 @EnableWebSecurity
 public class WebSecurityConfig {
 
-	@Autowired
-	private RememberMeServices rememberMeServices;
+    @Autowired
+    private RememberMeServices rememberMeServices;
 
-	@Bean
-	public PasswordEncoder passwordEncoder() {
-		return new BCryptPasswordEncoder();
-	}
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
 
-	@Bean
-	public AuthenticationManager authenticationManager(List<AuthenticationProvider> authenticationProviders) {
-		return new ProviderManager(authenticationProviders);
-	}
+    @Bean
+    public AuthenticationManager authenticationManager(List<AuthenticationProvider> authenticationProviders) {
+        return new ProviderManager(authenticationProviders);
+    }
 
-	@Bean
-	public SecurityContextRepository securityContextRepository() {
-		return new DelegatingSecurityContextRepository(
-				new RequestAttributeSecurityContextRepository(),
-				new HttpSessionSecurityContextRepository()
-		);
-	}
+    @Bean
+    public SecurityContextRepository securityContextRepository() {
+        return new DelegatingSecurityContextRepository(new RequestAttributeSecurityContextRepository(),
+                new HttpSessionSecurityContextRepository());
+    }
 
-	@Bean
-	public SecurityFilterChain filterChain(HttpSecurity http, SecurityContextRepository securityContextRepository) throws Exception {
-		// Pokud se tady nechá odkaz na login, tak se Vaadin nedostane na
-		// widgetset na login page (místo JSON response přijde login page...)
-		// Všechny možné ostatní postupy selhaly, takže to je nakonec udělané
-		// tak, že se přihlašování udělá programaticky. To ale nefunguje bez
-		// springSecurityFilterChain ... a ten zase chce http definici ... a ta
-		// chce aspoň jeden vstupní bod.
-		http.authorizeRequests().requestMatchers(AntPathRequestMatcher.antMatcher("/**")).permitAll();
+    @Bean
+    public SecurityFilterChain filterChain(HttpSecurity http, SecurityContextRepository securityContextRepository)
+            throws Exception {
+        // Pokud se tady nechá odkaz na login, tak se Vaadin nedostane na
+        // widgetset na login page (místo JSON response přijde login page...)
+        // Všechny možné ostatní postupy selhaly, takže to je nakonec udělané
+        // tak, že se přihlašování udělá programaticky. To ale nefunguje bez
+        // springSecurityFilterChain ... a ten zase chce http definici ... a ta
+        // chce aspoň jeden vstupní bod.
+        http.authorizeHttpRequests(c -> c.requestMatchers("/**").permitAll());
 		/*http.logout(c -> c.logoutUrl("/logout").logoutSuccessUrl("/").deleteCookies("JSESSIONID").deleteCookies(
 				"remember-me").invalidateHttpSession(true));*/
-		// https://docs.spring.io/spring-security/reference/5.8/migration/servlet/session-management.html#_require_explicit_saving_of_securitycontextrepository
-		// https://stackoverflow.com/questions/75618616/auto-login-after-registration-spring-boot-3-spring-security-6
-		http.securityContext((securityContext) -> securityContext.requireExplicitSave(true).securityContextRepository(securityContextRepository));
-		// https://www.mkyong.com/spring-security/spring-security-remember-me-example/
-		http.rememberMe(c -> c.rememberMeServices(rememberMeServices).key(SecurityConfig.REMEMBER_ME_KEY));
-		http.csrf(c -> c.disable());
+        // https://docs.spring.io/spring-security/reference/5.8/migration/servlet/session-management.html#_require_explicit_saving_of_securitycontextrepository
+        // https://stackoverflow.com/questions/75618616/auto-login-after-registration-spring-boot-3-spring-security-6
+        http.securityContext((securityContext) -> securityContext.requireExplicitSave(true)
+                .securityContextRepository(securityContextRepository));
+        // https://www.mkyong.com/spring-security/spring-security-remember-me-example/
+        http.rememberMe(c -> c.rememberMeServices(rememberMeServices).key(SecurityConfig.REMEMBER_ME_KEY));
+        http.csrf(c -> c.disable());
 
-		return http.build();
-	}
+        return http.build();
+    }
 }
