@@ -5,7 +5,7 @@ import cz.gattserver.grass.core.services.VersionInfoService;
 import cz.gattserver.grass.core.services.impl.MailServiceImpl;
 import cz.gattserver.grass.core.services.impl.VersionInfoServiceImpl;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.web.servlet.ServletComponentScan;
+import org.springframework.boot.web.server.servlet.context.ServletComponentScan;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
 import org.springframework.context.annotation.PropertySource;
