@@ -1,19 +1,19 @@
 package cz.gattserver.grass.monitor.processor.item;
 
-import elemental.json.JsonObject;
+import tools.jackson.databind.JsonNode;
 
 public class DiskStatusPartItemTO extends ListPartItemTO<DiskStatusMonitorItemTO> {
 
-	public DiskStatusPartItemTO() {
-	}
-	
-	public DiskStatusPartItemTO(JsonObject value) {
-		super(value);
-	}
+    public DiskStatusPartItemTO() {
+    }
 
-	@Override
-	protected DiskStatusMonitorItemTO createItem(JsonObject jsonObject) {
-		return new DiskStatusMonitorItemTO(jsonObject);
-	}
+    public DiskStatusPartItemTO(JsonNode value) {
+        super(value);
+    }
+
+    @Override
+    protected DiskStatusMonitorItemTO createItem(JsonNode jsonObject) {
+        return new DiskStatusMonitorItemTO(jsonObject);
+    }
 
 }
