@@ -1,55 +1,56 @@
 package cz.gattserver.grass.monitor.processor.item;
 
-import elemental.json.JsonObject;
+
+import tools.jackson.databind.JsonNode;
 
 public class JVMMemoryMonitorItemTO extends MonitorItemTO {
 
-	private long usedMemory;
-	private long freeMemory;
-	private long maxMemory;
-	private long totalMemory;
+    private long usedMemory;
+    private long freeMemory;
+    private long maxMemory;
+    private long totalMemory;
 
-	public JVMMemoryMonitorItemTO() {
-	}
+    public JVMMemoryMonitorItemTO() {
+    }
 
-	public JVMMemoryMonitorItemTO(JsonObject jsonObject) {
-		super(jsonObject);
-		usedMemory = (long) jsonObject.getNumber("usedMemory");
-		freeMemory = (long) jsonObject.getNumber("freeMemory");
-		maxMemory = (long) jsonObject.getNumber("maxMemory");
-		totalMemory = (long) jsonObject.getNumber("totalMemory");
-	}
+    public JVMMemoryMonitorItemTO(JsonNode jsonObject) {
+        super(jsonObject);
+        usedMemory = jsonObject.get("usedMemory").asLong();
+        freeMemory = jsonObject.get("freeMemory").asLong();
+        maxMemory = jsonObject.get("maxMemory").asLong();
+        totalMemory = jsonObject.get("totalMemory").asLong();
+    }
 
-	public long getUsedMemory() {
-		return usedMemory;
-	}
+    public long getUsedMemory() {
+        return usedMemory;
+    }
 
-	public void setUsedMemory(long usedMemory) {
-		this.usedMemory = usedMemory;
-	}
+    public void setUsedMemory(long usedMemory) {
+        this.usedMemory = usedMemory;
+    }
 
-	public long getFreeMemory() {
-		return freeMemory;
-	}
+    public long getFreeMemory() {
+        return freeMemory;
+    }
 
-	public void setFreeMemory(long freeMemory) {
-		this.freeMemory = freeMemory;
-	}
+    public void setFreeMemory(long freeMemory) {
+        this.freeMemory = freeMemory;
+    }
 
-	public long getMaxMemory() {
-		return maxMemory;
-	}
+    public long getMaxMemory() {
+        return maxMemory;
+    }
 
-	public void setMaxMemory(long maxMemory) {
-		this.maxMemory = maxMemory;
-	}
+    public void setMaxMemory(long maxMemory) {
+        this.maxMemory = maxMemory;
+    }
 
-	public long getTotalMemory() {
-		return totalMemory;
-	}
+    public long getTotalMemory() {
+        return totalMemory;
+    }
 
-	public void setTotalMemory(long totalMemory) {
-		this.totalMemory = totalMemory;
-	}
+    public void setTotalMemory(long totalMemory) {
+        this.totalMemory = totalMemory;
+    }
 
 }
