@@ -1,33 +1,32 @@
 package cz.gattserver.grass.monitor.processor.item;
 
-import elemental.json.JsonObject;
+import tools.jackson.databind.JsonNode;
 
 public class SMARTMonitorItemTO extends MonitorItemTO {
 
-	private String time;
+    private String time;
 
-	public SMARTMonitorItemTO() {
-	}
+    public SMARTMonitorItemTO() {
+    }
 
-	public SMARTMonitorItemTO(String time, String stateDetails) {
-		this.time = time;
-		this.stateDetails = stateDetails;
-	}
+    public SMARTMonitorItemTO(String time, String stateDetails) {
+        this.time = time;
+        this.stateDetails = stateDetails;
+    }
 
-	public SMARTMonitorItemTO(JsonObject jsonObject) {
-		super(jsonObject);
-		if (monitorState == MonitorState.ERROR)
-			return;
-		time = jsonObject.getString("time");
-		stateDetails = jsonObject.getString("stateDetails");
-	}
+    public SMARTMonitorItemTO(JsonNode jsonObject) {
+        super(jsonObject);
+        if (monitorState == MonitorState.ERROR) return;
+        time = jsonObject.get("time").asText();
+        stateDetails = jsonObject.get("stateDetails").asText();
+    }
 
-	public String getTime() {
-		return time;
-	}
+    public String getTime() {
+        return time;
+    }
 
-	public void setTime(String time) {
-		this.time = time;
-	}
+    public void setTime(String time) {
+        this.time = time;
+    }
 
 }
