@@ -1,18 +1,18 @@
 package cz.gattserver.grass.monitor.processor.item;
 
-import elemental.json.JsonObject;
+import tools.jackson.databind.JsonNode;
 
 public class ServersPartItemTO extends ListPartItemTO<URLMonitorItemTO> {
 
 	public ServersPartItemTO() {
 	}
 
-	public ServersPartItemTO(JsonObject jsonObject) {
+	public ServersPartItemTO(JsonNode jsonObject) {
 		super(jsonObject);
 	}
 
 	@Override
-	protected URLMonitorItemTO createItem(JsonObject jsonObject) {
+	protected URLMonitorItemTO createItem(JsonNode jsonObject) {
 		return new URLMonitorItemTO(jsonObject);
 	}
 
