@@ -1,19 +1,18 @@
 package cz.gattserver.grass.monitor.processor.item;
 
-import elemental.json.JsonObject;
+import tools.jackson.databind.JsonNode;
 
 public class ServicesPartItemTO extends ListPartItemTO<ServicesMonitorItemTO> {
 
 	public ServicesPartItemTO() {
 	}
 
-	public ServicesPartItemTO(JsonObject jsonObject) {
+	public ServicesPartItemTO(JsonNode jsonObject) {
 		super(jsonObject);
 	}
 
 	@Override
-	protected ServicesMonitorItemTO createItem(JsonObject jsonObject) {
+	protected ServicesMonitorItemTO createItem(JsonNode jsonObject) {
 		return new ServicesMonitorItemTO(jsonObject);
 	}
-
 }
